refactor: Mapper 패키지 Util로 변경
